feat(task): 각 도메인에 JpaRepository 생성
