+++ task/src/main/java/com/kimtaehoonki/task/project/domain/ProjectRepository.java
@@ -1,0 +1,7 @@
+package com.kimtaehoonki.task.project.domain;
+
+import com.kimtaehoonki.task.project.domain.entity.Project;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProjectRepository extends JpaRepository<Project, Long> {
+}
