+++ task/src/main/java/com/kimtaehoonki/task/tag/domain/TagRepository.java
@@ -1,0 +1,7 @@
+package com.kimtaehoonki.task.tag.domain;
+
+import com.kimtaehoonki.task.project.domain.entity.Project;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TagRepository extends JpaRepository<Tag, Long> {
+}
